httputil with dyn headers
